nlp_training/New folder/requirements.txt
pandas>=2.2.0
polars>=1.8.0
numpy>=1.26.0
scikit-learn>=1.4.0
scipy>=1.11.0
nltk>=3.8.1
spacy>=3.7.0
sentence-transformers>=3.0.0
faiss-cpu>=1.8.0
transformers>=4.41.0,<5
torch>=2.2.0
fastapi>=0.111.0
uvicorn>=0.30.0
joblib>=1.4.0
pyarrow>=15.0.0
matplotlib>=3.8.0
seaborn>=0.13.0
requests>=2.32.0
pypdf>=5.4.0
python-multipart>=0.0.20
langchain>=0.2.16
langchain-community>=0.2.16
langchain-openai>=0.1.24
langchain-google-genai>=2.0.0
